Colors, strings and themes xml files moved to coreUi module
