+++ app/src/androidTest/java/com/klewerro/mitemperature2alt/presentation/addThermometer/search/SearchThermometersRobot.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
 import com.klewerro.mitemperature2alt.MainActivityComposeRule
-import com.klewerro.mitemperature2alt.R
+import com.klewerro.mitemperature2alt.coreUi.R
 
 @OptIn(ExperimentalTestApi::class)
 class SearchThermometersRobot(
