+++ app/src/main/java/com/klewerro/mitemperature2alt/presentation/addThermometer/ConnectThermometerViewModel.kt
@@ -3,7 +3,7 @@ package com.klewerro.mitemperature2alt.presentation.addThermometer
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.klewerro.mitemperature2alt.R
+import com.klewerro.mitemperature2alt.coreUi.R
 import com.klewerro.mitemperature2alt.domain.usecase.thermometer.connect.ConnectToDeviceUseCase
 import com.klewerro.mitemperature2alt.domain.usecase.thermometer.operations.ReadCurrentThermometerStatusUseCase
 import com.klewerro.mitemperature2alt.domain.usecase.thermometer.persistence.SaveThermometerUseCase
