+++ app/src/main/java/com/klewerro/mitemperature2alt/presentation/mainscreen/TopBar.kt
@@ -15,8 +15,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.klewerro.mitemperature2alt.R
 import com.klewerro.mitemperature2alt.coreUi.LocalSpacing
+import com.klewerro.mitemperature2alt.coreUi.R
 
 @Composable
 fun TopBar(title: String, shouldBeButtonVisible: Boolean, onButtonClick: () -> Unit) {
