+++ coreUi/src/main/res/values/colors.xml
