+++ coreUi/src/main/res/values/strings.xml
